requirements.txt
requests
python-dotenv
Flask
Flask-Limiter